refactor: decompose session detail into smaller components and hook
Extract from app/session/detail/[id].tsx (731 → 138 lines):
- hooks/useSessionDetail.ts: all state, effects, and callbacks
- components/session/detail/SummaryCard.tsx: date/duration/sets/volume card
- components/session/detail/RatingNotesCard.tsx: rating widget + notes section
- components/session/detail/ExerciseGroupRow.tsx: exercise/set rendering
- components/session/detail/TemplateModal.tsx: save-as-template modal

Also exports ThemeColors type from useThemeColors for component props.
